move cw into separate config
