+++ code/parachain/runtime/picasso/src/contracts.rs
@@ -1,0 +1,168 @@
+
+use ::cosmwasm::pallet_hook::PalletHook;
+use sp_core::ConstU32;
+use cosmwasm::instrument::CostRules;
+
+
+use super::*;
+
+/// Native <-> Cosmwasm account mapping
+pub struct AccountToAddr;
+
+impl Convert<alloc::string::String, Result<AccountId, ()>> for AccountToAddr {
+	fn convert(a: alloc::string::String) -> Result<AccountId, ()> {
+		let account =
+			ibc_primitives::runtime_interface::ss58_to_account_id_32(&a).map_err(|_| ())?;
+		Ok(account.into())
+	}
+}
+
+impl Convert<AccountId, alloc::string::String> for AccountToAddr {
+	fn convert(a: AccountId) -> alloc::string::String {
+		let account = ibc_primitives::runtime_interface::account_id_to_ss58(a.into(), 49);
+		String::from_utf8_lossy(account.as_slice()).to_string()
+	}
+}
+
+impl Convert<Vec<u8>, Result<AccountId, ()>> for AccountToAddr {
+	fn convert(a: Vec<u8>) -> Result<AccountId, ()> {
+		Ok(<[u8; 32]>::try_from(a).map_err(|_| ())?.into())
+	}
+}
+
+/// Native <-> Cosmwasm asset mapping
+pub struct AssetToDenom;
+
+impl Convert<alloc::string::String, Result<CurrencyId, ()>> for AssetToDenom {
+	fn convert(currency_id: alloc::string::String) -> Result<CurrencyId, ()> {
+		core::str::FromStr::from_str(&currency_id).map_err(|_| ())
+	}
+}
+
+impl Convert<CurrencyId, alloc::string::String> for AssetToDenom {
+	fn convert(CurrencyId(currency_id): CurrencyId) -> alloc::string::String {
+		alloc::format!("{}", currency_id)
+	}
+}
+
+
+parameter_types! {
+	pub const CosmwasmPalletId: PalletId = PalletId(*b"cosmwasm");
+	pub const ChainId: &'static str = "composable-network-picasso";
+	pub const MaxInstrumentedCodeSize: u32 = 1024 * 1024;
+	pub const MaxContractLabelSize: u32 = 64;
+	pub const MaxContractTrieIdSize: u32 = Hash::len_bytes() as u32;
+	pub const MaxInstantiateSaltSize: u32 = 128;
+	pub const MaxFundsAssets: u32 = 32;
+	pub const CodeTableSizeLimit: u32 = 4096;
+	pub const CodeGlobalVariableLimit: u32 = 256;
+	pub const CodeParameterLimit: u32 = 128;
+	pub const CodeBranchTableSizeLimit: u32 = 256;
+
+	// TODO: benchmark for proper values
+	pub const CodeStorageByteDeposit: u32 = 1_000_000;
+	pub const ContractStorageByteReadPrice: u32 = 1;
+	pub const ContractStorageByteWritePrice: u32 = 1;
+	pub WasmCostRules: CostRules<Runtime> = Default::default();
+}
+
+impl cosmwasm::Config for Runtime {
+	type RuntimeEvent = RuntimeEvent;
+	type AccountIdExtended = AccountId;
+	type PalletId = CosmwasmPalletId;
+	type MaxFrames = ConstU32<64>;
+	type MaxCodeSize = ConstU32<{ 512 * 1024 }>;
+	type MaxInstrumentedCodeSize = MaxInstrumentedCodeSize;
+
+	#[cfg(feature = "testnet")]
+	type MaxMessageSize = ConstU32<{ 128 * 1024 }>;
+
+	#[cfg(not(feature = "testnet"))]
+	type MaxMessageSize = ConstU32<{ 32 * 1024 }>;
+
+	type AccountToAddr = AccountToAddr;
+
+	type AssetToDenom = AssetToDenom;
+
+	type Balance = Balance;
+	type AssetId = CurrencyId;
+	type Assets = AssetsTransactorRouter;
+	type NativeAsset = Balances;
+	type ChainId = ChainId;
+	type MaxContractLabelSize = MaxContractLabelSize;
+	type MaxContractTrieIdSize = MaxContractTrieIdSize;
+	type MaxInstantiateSaltSize = MaxInstantiateSaltSize;
+	type MaxFundsAssets = MaxFundsAssets;
+
+	type CodeTableSizeLimit = CodeTableSizeLimit;
+	type CodeGlobalVariableLimit = CodeGlobalVariableLimit;
+	type CodeStackLimit = ConstU32<{ u32::MAX }>;
+
+	type CodeParameterLimit = CodeParameterLimit;
+	type CodeBranchTableSizeLimit = CodeBranchTableSizeLimit;
+	type CodeStorageByteDeposit = CodeStorageByteDeposit;
+	type ContractStorageByteReadPrice = ContractStorageByteReadPrice;
+	type ContractStorageByteWritePrice = ContractStorageByteWritePrice;
+
+	type WasmCostRules = WasmCostRules;
+	type UnixTime = Timestamp;
+	type WeightInfo = cosmwasm::weights::SubstrateWeight<Runtime>;
+	type IbcRelayerAccount = TreasuryAccount;
+
+	type IbcRelayer = cosmwasm::NoRelayer<Runtime>;
+
+	type PalletHook = ();
+
+	#[cfg(feature = "testnet")]
+	type UploadWasmOrigin = system::EnsureSigned<Self::AccountId>;
+
+	#[cfg(feature = "testnet")]
+	type ExecuteWasmOrigin = system::EnsureSigned<Self::AccountId>;
+
+	// really need to do EnsureOnOf<Sudo::key, >
+	#[cfg(not(feature = "testnet"))]
+	type UploadWasmOrigin = system::EnsureSignedBy<TechnicalCommitteeMembership, Self::AccountId>;
+
+	#[cfg(not(feature = "testnet"))]
+	type ExecuteWasmOrigin = frame_support::traits::EitherOfDiverse<
+		system::EnsureSignedBy<TechnicalCommitteeMembership, Self::AccountId>,
+		system::EnsureSignedBy<ReleaseMembership, Self::AccountId>,
+	>;
+}
+
+pub struct Precompiles;
+
+
+// impl PalletHook<Runtime> for Precompiles {
+//     fn info(
+// 		contract_address: &::cosmwasm::types::AccountIdOf<Runtime>,
+// 	) -> Option<::cosmwasm::types::PalletContractCodeInfo<::cosmwasm::types::AccountIdOf<Runtime>, ::cosmwasm::types::ContractLabelOf<Runtime>, ::cosmwasm::types::ContractTrieIdOf<Runtime>>> {
+//         todo!()
+//     }
+
+//     fn execute<'a>(
+// 		vm: &mut OwnedWasmiVM<::cosmwasm::runtimes::vm::CosmwasmVM<'a, Runtime>>,
+// 		entrypoint: ::cosmwasm::types::EntryPoint,
+// 		message: &[u8],
+// 	) -> Result<
+// 		ContractResult<Response<<OwnedWasmiVM<::cosmwasm::runtimes::vm::CosmwasmVM<'a, Runtime>> as VMBase>::MessageCustom>>,
+// 		VmErrorOf<OwnedWasmiVM<::cosmwasm::runtimes::vm::CosmwasmVM<'a, Runtime>>>,
+// 	> {
+//         todo!()
+//     }
+
+//     fn run<'a>(
+// 		vm: &mut OwnedWasmiVM<::cosmwasm::runtimes::vm::CosmwasmVM<'a, Runtime>>,
+// 		entrypoint: ::cosmwasm::types::EntryPoint,
+// 		message: &[u8],
+// 	) -> Result<Vec<u8>, VmErrorOf<OwnedWasmiVM<::cosmwasm::runtimes::vm::CosmwasmVM<'a, Runtime>>>> {
+//         todo!()
+//     }
+
+//     fn query<'a>(
+// 		vm: &mut OwnedWasmiVM<::cosmwasm::runtimes::vm::CosmwasmVM<'a, Runtime>>,
+// 		message: &[u8],
+// 	) -> Result<ContractResult<QueryResponse>, VmErrorOf<OwnedWasmiVM<::cosmwasm::runtimes::vm::CosmwasmVM<'a, Runtime>>>> {
+//         todo!()
+//     }
+// }
